@@ -3,52 +3,51 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^1.1.0",
-    "@material-ui/icons": "^1.0.0-beta.43",
-    "ajv-keywords": "^3.1.0",
-    "antd": "^3.6.2",
+    "@material-ui/core": "^1.2.3",
+    "@material-ui/icons": "^1.1.0",
+    "ajv": "^6.5.1",
+    "ajv-keywords": "^3.2.0",
+    "antd": "^3.6.3",
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
-    "classnames": "^2.2.6",
     "draft-js": "^0.10.5",
     "draft-js-buttons": "^2.0.1",
-    "draft-js-plugins-editor": "^2.0.8",
+    "draft-js-plugins-editor": "^2.1.1",
     "draft-js-static-toolbar-plugin": "^2.0.2",
     "draftjs-to-html": "^0.8.4",
+    "echarts": "^4.1.0",
     "install": "^0.11.0",
-    "material-ui": "^1.0.0-beta.44",
-    "npm": "^6.0.0",
-    "prop-types": "^15.6.1",
-    "rc-calendar": "^9.6.1",
-    "react": "^16.2.0",
-    "react-avatar-editor": "^11.0.3",
-    "react-date-picker": "^6.10.1",
-    "react-dom": "^16.2.0",
+    "isomorphic-fetch": "^2.2.1",
+    "jwt-decode": "^2.2.0",
+    "material-ui": "^1.0.0-beta.47",
+    "material-ui-animated-fab": "^1.0.3",
+    "npm": "^6.1.0",
+    "prop-types": "^15.6.2",
+    "qs": "^6.5.2",
+    "rc-calendar": "^9.6.2",
+    "react": "^16.4.1",
+    "react-avatar-editor": "^11.0.4",
+    "react-date-picker": "^6.10.4",
+    "react-dom": "^16.4.1",
     "react-draft-wysiwyg": "^1.12.13",
     "react-dropzone": "^4.2.11",
+    "react-helmet": "^5.2.0",
     "react-html-parser": "^2.0.2",
-    "react-quill": "^1.2.7",
+    "react-quill": "^1.3.0",
     "react-redux": "^5.0.7",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^4.0.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-router-redux": "^4.0.8",
     "react-scripts": "1.1.1",
     "redux": "^3.7.2",
-    "redux-persist": "^5.9.1",
-    "redux-persist-transform-filter": "0.0.17",
-    "redux-thunk": "^2.2.0",
-    "superagent": "^3.8.3"
+    "redux-persist": "^5.10.0",
+    "redux-persist-transform-filter": "0.0.18",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "babel-plugin-transform-decorators": "^6.24.1",
-    "jwt-decode": "^2.2.0",
-    "react-helmet": "^5.2.0",
-    "react-tap-event-plugin": "^3.0.2"
   }
 }
