@@ -1,13 +1,17 @@
 import React, {Component} from 'react';
 import {connect} from "react-redux"
-import {Route, Switch} from "react-router-dom"
+import {Route, Switch,Link} from "react-router-dom"
+import BZ from "./ButtonZoom"
+import {Button,
+    List,
+    ListItem,
+    Paper,
+    ListItemIcon,
+    Divider,
+    ListItemText,
+    ListItemSecondaryAction
 
-import Button from 'material-ui/Button';
-import List from 'material-ui/List';
-import ListItem from 'material-ui/List';
-
-
-import Paper from 'material-ui/Paper'
+} from 'material-ui';
 
 
 
@@ -16,82 +20,59 @@ import PaperManage from '../PaperManage'
 import GradeStatistics from '../GradeStatistics'
 import PaperGenerate from '../PaperGenerate'
 import PaperView from '../PaperView'
-
+import Bar from '../../../../../top/components/Bar'
 import {getCourseList} from "./actions";
-
+import {
+    Home, Search as SearchIcon, Message as MessageIcon,
+    Announcement as AnnouncementIcon,
+    Extension as ExtensionIcon
+} from "@material-ui/icons/es/index"
 const menuStyle = {
     display: 'inline-block',
     width: "20%",
-    height: "100%",
+    height: "80%",
     float: "left",
     margin: "20px"
 };
 
 const windowStyle = {
     display: 'inline-block',
-    width: "70%",
-    height: "100%",
+    width: "98%",
+    height: "80%",
     float: "left",
     margin: "20px"
 };
 
 class TeacherMain extends Component {
 
-    componentWillMount() {
+    componentDidMount() {
         this.props.getCourseList(0, 0);
     }
 
     render() {
-        const {match, course_list} = this.props;
+        const {token, match, course_list} = this.props;
         // console.log('list', courseList);
         const courseListItems = course_list.map(
             (courseInfo, index)=>{
                 return (
+                    <div>
                     <ListItem key={ index} >
-                        <p>
-                            {courseInfo.course_name}
-                        </p>
-                        <Button
-                            onClick={() => {
-                                const path = `${match.url}/question_manage/${courseInfo.course_id}`;
-                                this.props.history.push(path);
-                            }}
-                        >
-                            题库
-                        </Button
-
-                        >
-                        <Button
-                            onClick={() => {
-                                const path = `${match.url}/paper_manage/${courseInfo.course_id}`;
-                                this.props.history.push(path);
-                            }}
-
-                        > 试卷</Button>
-                        <Button
-                            onClick={() => {
-                                const path = `${match.url}/grade_statistics/${courseInfo.course_id}`;
-                                this.props.history.push(path);
-                            }}
-
-                        > 统计</Button>
+                            <BZ courseInfo={courseInfo} courseName={courseInfo.course_name} course_id={courseInfo.course_id} history={this.props.history} match={match} />
                     </ListItem>
+                    </div>
                 )
             }
 
-        )
+        );
+
 
         return (
+            <Bar listItems={courseListItems}>
             <div>
-               <Paper style={menuStyle}>
-                   <List>
-                       {courseListItems}
-                   </List>
-               </Paper>
                 <Paper style={windowStyle}>
                     <Route path={`${match.url}/question_manage/:course_id`} component={QuestionManage}/>
                     <Route exact path={`${match.url}/paper_manage/:course_id`} component={PaperManage}/>
-                    <Route path={`${match.url}/paper_manage/:course_id/paper_generate/`} component={PaperGenerate}/>
+                    <Route path={`${match.url}/paper_manage/:course_id/paper_generate`} component={PaperGenerate}/>
                     <Route path={`${match.url}/paper_manage/:course_id/paper_view/:paper_id`} component={PaperView}/>
                     <Route path={`${match.url}/grade_statistics/:course_id`} component={GradeStatistics}/>
 
@@ -99,6 +80,7 @@ class TeacherMain extends Component {
 
                 </Paper>
             </div>
+            </Bar>
         );
     }
 }
@@ -106,6 +88,7 @@ class TeacherMain extends Component {
 
 const mapStateToProps = (state) => ({
     course_list: state.online_testing.teacher_main.course_list,
+    token: state.online_testing.teacher_main.token,
 });
 
 const mapDispatchToProps = (dispatch) => {
