@@ -9,9 +9,10 @@ from rest_framework.response import Response
 from rest_framework import viewsets, status
 from rest_framework.decorators import action
 import numpy as np
+from rest_framework.views import APIView
 from rest_framework.viewsets import GenericViewSet, mixins
 
-from authentication.models import Student, Course
+from authentication.models import Student, Course, Faculty
 from online_testing.models import Question, Paper, Examination
 from online_testing.serializers import QuestionSerializer, QuestionDetailSerializer, \
     PaperSerializer, ExaminationSerializer, PaperDetailSerializer
@@ -19,6 +20,13 @@ from online_testing.filters import QuestionFilter
 from online_testing.permissions import ExamInfoAccessPermission, QuestionPermission, PaperPermission
 from rest_framework import permissions
 
+class CourseQuery(APIView):
+    def get(self, request):
+        if request.user.user_type == 1:
+            return Response(Course.objects.all().values_list('course_id', 'name'))
+        elif request.user.user_type == 2:
+            faculty = Faculty.objects.all().get(username=request.user.username)
+            return Response(faculty.teacher_course.all().values_list('course_id', 'name'))
 
 class QuestionViewSet(viewsets.ModelViewSet):
     # not safe: update(object), partial_update(object), delete(object), insert,
@@ -219,6 +227,7 @@ class PaperViewSet(mixins.CreateModelMixin,
                         paper['done'] = False
                 except ObjectDoesNotExist:
                     paper['done'] = False
+                print(paper)
         return Response({'paper_list': data})
 
 
@@ -365,7 +374,10 @@ class AnalysisViewSet(GenericViewSet):
     def studentGradeList(self, request):
         data = []
         map = {}
+        course_id = request.query_params.get('course_id')
         for exam in Examination.objects.all().filter(student=request.user.username):
+            if exam.paper.course.course_id != course_id:
+                continue
             s = PaperDetailSerializer(exam.paper)
             if map.get(exam.paper) is None:
                 map[exam.paper] = Examination.objects.all().\
@@ -392,6 +404,7 @@ class AnalysisViewSet(GenericViewSet):
             }
             for exam in Examination.objects.all().filter(paper=paper):
                 d['whoTakeThisTest'].append({
+                    'studentID': exam.student.username.username,
                     'studentName': exam.student.name,
                     'studentScore': exam.score
                 })
@@ -401,6 +414,7 @@ class AnalysisViewSet(GenericViewSet):
     @action(methods=['get'], detail=False)
     def studentList(self, request):
         data = []
+        # should change
         for student in Student.objects.all():
             d = {
                 'studentID': student.username.username,
@@ -459,13 +473,17 @@ class AnalysisViewSet(GenericViewSet):
                 'testName': paper.paper_name,
                 'content': []
             }
-            for i in range(0, 5):
+            for i, question in enumerate(paper.question_id_list.all()):
+                if i > 10:
+                    break
                 d['content'].append({
-                    'questionID': int(np.random.randint(100, 300)),
+                    'questionID': question.question_id,
                     'answerRate': '%d%%' % int(np.random.randint(0, 100))
                 })
-            data['multiChoices'].append(d)
-            data['judge'].append(d)
+                if question.type == 'Choice':
+                    data['multiChoices'].append(d)
+                else:
+                    data['judge'].append(d)
         return Response(data)
 
 
