@@ -3,12 +3,7 @@ import {connect} from 'react-redux';
 import {Route, Switch} from "react-router-dom"
 import StudentMain from './views/student/StudentMain'
 import TeacherMain from './views/teacher/TeacherMain'
-
-import Button from 'material-ui/Button';
-import AddIcon from '@material-ui/icons/Add';
-import Icon from 'material-ui/Icon';
-import DeleteIcon from '@material-ui/icons/Delete';
-
+import {Button} from 'material-ui'
 
 const styles = {
     backgroundColor: '#f0f0ee',
@@ -18,15 +13,18 @@ const styles = {
 };
 
 
-let Identity = "teacher";
-
-
 
 
 class OnlineTesting extends Component {
+
+    state={
+    };
+
+
     render() {
         const {match} = this.props;
-        const routePath= Identity === "teacher" ? (
+        console.log(localStorage.getItem('type'));
+        const routePath= (localStorage.getItem('type') === "2") ? (
             <Route path={`${match.url}`} component={TeacherMain}/>
         ):(
             <Route path={`${match.url}`} component={StudentMain}/>
