@@ -1,6 +1,5 @@
 import {
-    ADD_PROBLEM_TEACHER,
-    DELETE_PROBLEMS_TEACHER,
+    CHANGE_PROBLEM_SHOW_LIST,
     GET_PROBLEM_LIST,
     GET_TEACHER_AND_TAG_LIST
 } from "./actions";
@@ -9,6 +8,7 @@ const initState = {
     tag_list: [],
     teacher_list:[],
     problem_list:[],
+    problem_should_show:[]
 };
 
 export function questionManageReducer(state = initState, action) {
@@ -17,18 +17,17 @@ export function questionManageReducer(state = initState, action) {
             return Object.assign({}, state, {
                 teacher_list: action.teacher_list,
                 tag_list: action.tag_list,
+                problem_list:[],
+                problem_should_show:[]
             });
         case GET_PROBLEM_LIST:
             return Object.assign({}, state, {
                 problem_list: action.problem_list,
+                problem_should_show: action.problem_should_show
             });
-        case ADD_PROBLEM_TEACHER:
+        case CHANGE_PROBLEM_SHOW_LIST:
             return Object.assign({}, state, {
-                problem_list:action.problem_list
-            });
-        case DELETE_PROBLEMS_TEACHER:
-            return Object.assign({}, state, {
-                problem_list:action.problem_list
+                problem_should_show: action.problem_should_show
             });
         default:
             return state;
